+++ contracts/AppConfig.sol
@@ -5,15 +5,15 @@ pragma solidity ^0.8.4;
  * Configuration functions for an App.
  */
 abstract contract AppConfig {
-    event ResourceChanged(uint32 indexed id, uint32 indexed weight, OP indexed op);
-    event ResourcePending(uint32 indexed id, uint32 indexed newWeight, OP indexed op);
+    event ResourceChanged(uint32 indexed id, uint256 indexed weight, OP indexed op);
+    event ResourcePending(uint32 indexed id, uint256 indexed newWeight, OP indexed op);
     struct ConfigEntry {
         string resourceId;
-        uint32 weight;
+        uint256 weight;
         uint32 index; // index in indexArray
         // pending action
         OP pendingOP;
-        uint32 pendingWeight;
+        uint256 pendingWeight;
         /* when the pending action was submitted */
         uint submitSeconds;
         uint256 requestTimes;
@@ -42,7 +42,7 @@ abstract contract AppConfig {
     struct ConfigRequest {
         uint32 id;
         string resourceId;
-        uint32 weight;
+        uint256 weight;
         /** Operation code for configuring resources, ADD 0; UPDATE: 1; DELETE: 2 */
         OP op;
     }
@@ -79,7 +79,7 @@ abstract contract AppConfig {
     function _configResource(ConfigRequest memory entry) internal {
         uint32 id = entry.id;
         string memory resourceId = entry.resourceId;
-        uint32 weight = entry.weight;
+        uint256 weight = entry.weight;
         OP op = entry.op;
         if (op == OP.ADD) {
             // id starts from `FIRST_CONFIG_ID`, if resourceId=>id > 0, it's added already.
@@ -128,7 +128,7 @@ abstract contract AppConfig {
         }
         emit ResourcePending(id, weight, op);
     }
-    function setPendingProp(uint32 id, OP op_, uint32 weight_) internal {
+    function setPendingProp(uint32 id, OP op_, uint256 weight_) internal {
         resourceConfigures[id].pendingOP = op_;
         resourceConfigures[id].pendingWeight = weight_;
         resourceConfigures[id].submitSeconds = block.timestamp;
@@ -153,7 +153,7 @@ abstract contract AppConfig {
                 continue;
             }
             OP op = config.pendingOP;
-            uint32 weight = config.pendingWeight;
+            uint256 weight = config.pendingWeight;
             if (op == OP.ADD) {
                 _mintConfig(address(this), id, weight, "add config");
                 config.weight = config.pendingWeight;
